Land one Gemini lane pin and seven coherence repairs (charter 4b): 5,958 pins in 1,264 rows
Lane harvest: codex luna pack r56_unserved1 0 of 12, Gemini pack
r56_unserved2 1 of 12 (dungeon/func_800BB144, one pin). Coherence
repairs: six pin-free rows and one pinned row (its keep erased) become
byte-exact at their module census recipe, seven recipe deviations
repaid (one leaving gcc 2.95.2, two dropping flags), all recorded in
ledger/recipe_trades.jsonl. Gates: 2 + 5 windows MATCH, SLUS MATCH,
row database and case ledger OK.

+++ src/dungeon/func_800942BC.c
@@ -71,7 +71,7 @@ s32 func_80099A1C(void *object, S_80099A1C_1 *position, S_80099A1C_2 *offset_dat
 
         {
             s32 half_offset;
-            u16 adjusted_coord;
+            u32 adjusted_coord;
 
             half_offset = ((S_80099A1C_3 *)object)->unk_00;
             adjusted_coord = ((S_80099A1C_0 *)scratch)->unk_B8;
